Add files via upload
Project Summary
Completed by: Nandana Pradeep 

This was an internship project on data science that I did for Afame technologies, where I was given 3 projects out of which I was asked to choose one and work on it. The project requirement was to analyze a dataset they had provided (I chose sales prediction) and make a sales prediction model using python. After downloading the dataset, I carefully studied its attributes and data and concluded that the data set had sales as their target variable/attribute, and it depended on Advertising through Television, Radio, and Newspaper. So, the unit of sales wasn’t specified but we could assume it in terms of thousands and the other attributes were the amount paid whose units were also to be assumed. I worked with python using Google Colab to develop the training model. Once I successfully created the model, I downloaded the PKL file of the training model. To simply show the working of my model I have created a small webpage using HTML, CSS, JavaScript, Flask, and Python. I have integrated the PKL file with the webpage and the user just need to enter the values and click on predict to get the predicted outcome. 
Git : https://github.com/Nandana-125/Afame-Technologies.git
Tools Used: Google colab, Python, JavaScript, CSS, Flask, Visual Studio